Diseño e Implementación de un Sistema de Clasificación Contable Automatizado para Facturas/src/Funciones Cloud/Model Training/requirements.txt
Flask==2.3.3
google-auth
google-cloud-storage
google-cloud-firestore
google-auth-oauthlib
functions-framework
pandas
numpy
PyJWT
google-cloud-secret-manager
joblib
scikit-learn
voluptuous
